moved createSlug to helper file & added regex to remove article words
