+++ modules/solutions/solutions.server.controller.js
@@ -15,7 +15,9 @@ let path = require('path'),
     votes = require('../votes/votes.server.controller'),
     proposals = require('../proposals/proposals.server.controller'),
     _ = require('lodash'),
-    seed = require('./seed/seed');
+    seed = require('./seed/seed'),
+    createSlug = require('../helpers/stuff');
+
 
 /**
  * Create a solution
@@ -316,12 +318,3 @@ exports.seedData = function (organizationId, issueId) {
     newSolution.save();
     return newSolution;
 };
-
-function createSlug(string) {
-    return string
-        .replace(/[^a-z0-9 -]/g, '') // remove invalid chars
-        .replace(/\s+/g, '-') // collapse whitespace and replace by -
-        .replace(/-+/g, '-') // collapse dashes
-        .replace(/^-+/, '') // trim - from start of text
-        .replace(/-+$/, '') // trim - from end of text
-}
